+++ src/Gateway.php
@@ -27,7 +27,7 @@ use Pronamic\WordPress\Pay\Payments\Payment;
  * Company: Pronamic
  *
  * @author  Remco Tolsma
- * @version 2.0.8
+ * @version 2.0.9
  * @since   1.1.0
  */
 class Gateway extends Core_Gateway {