+++ src/Integration.php
@@ -22,7 +22,7 @@ use WP_User;
  * Company: Pronamic
  *
  * @author  Remco Tolsma
- * @version 2.0.8
+ * @version 2.0.9
  * @since   1.0.0
  */
 class Integration extends AbstractIntegration {