+++ src/LocaleHelper.php
@@ -17,7 +17,7 @@ namespace Pronamic\WordPress\Pay\Gateways\Mollie;
  * Company: Pronamic
  *
  * @author  Remco Tolsma
- * @version 2.1.0
+ * @version 2.0.9
  * @since   1.0.0
  */
 class LocaleHelper {