+++ tests/src/MethodsTest.php
@@ -19,7 +19,7 @@ use Pronamic\WordPress\Pay\Core\PaymentMethods;
  * Company: Pronamic
  *
  * @author  Remco Tolsma
- * @version 2.1.0
+ * @version 2.0.9
  * @since   1.0.0
  */
 class MethodsTest extends \PHPUnit_Framework_TestCase {