+++ tests/src/PaymentRequestTest.php
@@ -16,7 +16,7 @@ use Pronamic\WordPress\Money\Money;
  * Payment request test
  *
  * @author  Remco Tolsma
- * @version 2.1.0
+ * @version 2.0.9
  * @since   1.0.0
  */
 class PaymentRequestTest extends \PHPUnit_Framework_TestCase {